refactor(F719): decompose skills-state.ts into natural locations
skills-state.ts was a grab-bag of unrelated functions — source scanning,
config writes, and staleness detection. Distribute each to its natural home:

- listSourceSkillNames, computeSourceManifestHash, checkStaleness → skill-source.ts
- readSkillsSyncState, writeSkillsSyncState, updateSkillMountPaths,
  removeCatCafeSkillCapabilities → skill-sync-config.ts

Delete skills-state.ts (210 lines). All 7 consumers updated.
101 tests pass, 0 regressions.
